Remove old CTE decoder
